vmm: refactor FD passing
As discussed in PR #7377, Cloud Hypervisor may potentially accept
different types of FDs on VM creation and/or resuming:
- FDs for net devices (currently supported);
- FDs for pmem devices for memory zones (WIP);
- FDs for storage devices, etc.

As FDs are passed as an array of anonymous numbers over UDS sockets,
there should be a way to assign them to the devices they belong to.

This PR introduces ExternalFdsConfig in vm_config.rs, which
matches, positionally, the list of device IDs/names to the list
of FDs.

Signed-off-by: Peter Oskolkov <posk@google.com>